work database
checked use homeController/Import method

+++ WeatherForecast/ViewModels/WeatherViewModel.cs
@@ -6,7 +6,7 @@ namespace WeatherForecast.ViewModels
     {
         public DateTime Date { get; set; }
         public double Temperature { get; set; }
-        public int RelativeHumidity { get; set; }
+        public double RelativeHumidity { get; set; }
         public double DewPoint { get; set; }
         public int AtmosphericPressure { get; set; }
         public string? WindDirection { get; set; }
